Remove YourKit profiler option
This option was not very useful as it didn't consider the difference
between warmups and measurements and wasn't aware of cleanup tasks.

In the long run we might want to replace our whole performance testing
infrastructure with the Gradle Profiler. Then we automatically support
all profilers that the Gradle Profiler supports. Until then JFR is a
good default and we won't have to maintain other profilers here.